Fix file detection as well as well as add tests
I wanted the program to error upon trying to search through a
folder and finding nothing (despite the folder itself existing).

Unfortunately, what I failed to realize is that everything in UNIX
is a file, including another folder. The os.listdir function will,
as a result, consider subfolders to count as "files".

This meant that a folder with nothing but subfolders will pass the
validation tests at program startup.

In the future, I may add a `--recursive` flag where in specifying
a folder, it will look for files in the root of that folder along
with its subfolders, but that is far down the line. For now I want
to keep things simple and make sure that a folder always has at
least one file.

Next steps: begin the parsing stage of questions/tokens